chatbot_api/requirements.txt
mysql-connector-python==9.1.0
langchain==0.3.8
langchain-openai==0.2.9
langchain_core==0.3.21
langchain_community==0.3.7
pydantic==2.10.0
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1